refactor(types): add types, refactor imports
issue -  #1147
